Rearrange classes into subpackages
Separate the classes for finding possible watch faces and generating the
code for the different times.
