+++ src/timepiece/TimeNamesGerman.java
